backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
websockets==12.0
pydantic==2.5.0
python-multipart==0.0.6
langchain==0.1.0
langgraph==0.0.62
langchain-openai==0.0.8
langchain-community==0.0.20
python-dotenv==1.0.0
httpx==0.25.2
aiofiles==23.2.1
neo4j==5.23.0
openai>=1.30.0
tavily-python>=0.3.5
neo4j-graphrag==1.10.0
